EC2 Return check + Slave delete
AWS CLI deletes now check the return JSON of each delete statement to evaluate whether a new machine has been deleted, whether it has been successfully deleted or if a machine was already in a terminated state and is still terminated. 

Also added a check so that slaves are matched and then safely deleted within both delete scripts.